+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
+## <next>
+* Add `dragItemZindex` prop for raising the SortableItems z-index on top of e.g. bootstrap Modal.
+* Added new example to show `dragItemZindex` in action.
 
 ## 1.0.0
 * Sorting with drag'n'drop